margin isolated trading grids api

+++ cryptos/models/binance/spot/plan.go
@@ -8,13 +8,13 @@ import (
 type Plan struct {
 	ID        string            `gorm:"size:20;primaryKey"`
 	Symbol    string            `gorm:"size:20;not null;uniqueIndex:unq_binance_spot_plans_symbol_timestamp"`
-	Side      int64             `gorm:"not null"`
+	Side      int               `gorm:"not null"`
 	Price     float64           `gorm:"not null"`
 	Quantity  float64           `gorm:"not null"`
 	Amount    float64           `gorm:"not null"`
 	Timestamp int64             `gorm:"not null;index;uniqueIndex:unq_binance_spot_plans_symbol_timestamp"`
 	Context   datatypes.JSONMap `gorm:"not null"`
-	Status    int64             `gorm:"not null"`
+	Status    int               `gorm:"not null"`
 	Remark    string            `gorm:"size:5000;not null"`
 	CreatedAt time.Time         `gorm:"not null;index"`
 	UpdatedAt time.Time         `gorm:"not null"`
